Bump R4J to version 2.6.0
Something is broken with how R4J was uploaded to jitpack so I had
to manually build the jar and add to the project. The only
complication was I also had to exclude sl4j from the jar in order
to not conflict with Spring.

I will replace the jar with 2.6.0 once that works again.

See https://github.com/stelar7/R4J/issues/119

This also coincides with the Riot API requirement to use puuid's
everywhere so there are also changes for that.

Lastly, there were some unused methods / params that got cleaned
up as well.